+++ CHANGELOG.md
@@ -7,7 +7,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/).
 
-## Unreleased
+## Unreleased - 2026-01-28
+
+### Changed
+
+- fix mapped tables did not use toFilter for searching leading to using non
+  mapped records without titles
+- add every type of column can now register it's value mapper
+- added parameter 'searchable' to define columns that can be searched on table
+
+## Unreleased - 2026-01-19
 
 ### Removed